+++ src/apollo.ts
@@ -1,15 +1,13 @@
 import { ApolloServer } from "apollo-server-express";
 
-import { typedefs } from "./graphql/schema";
-import { gqlResolvers } from "./graphql/resolvers";
+import { schema} from "./graphql/schema";
 
 import auth from "./services/auth";
 import events from "./services/events";
 import users from "./services/users";
 
 export const server = new ApolloServer({
-  typeDefs: typedefs,
-  resolvers: gqlResolvers,
+  schema,
   context: (req) => {
     return {
       ...req,
