+++ src/graphql/models/auth.ts
@@ -1,0 +1,8 @@
+import gql from "graphql-tag";
+
+export const typeDefs = gql`
+  type AuthPayload {
+    token: String
+    user: User
+  }
+`;
