+++ src/graphql/models/event.ts
@@ -1,0 +1,18 @@
+import gql from "graphql-tag";
+
+export const typeDefs = gql`
+  type Event {
+    id: Int!
+    title: String!
+    content: String!
+    location: String
+    published: Boolean!
+    host: Int!
+    createdAt: DateTime
+  }
+
+  type EventAttendee {
+    userId: Int!
+    eventId: Int!
+  }
+`;
